refactor: apply Pythonic naming conventions to module structure
- rename overviewpy.py → overview.py to avoid the overviewpy.overviewpy double-name
- rename datafilesummarizer/ → summarizer/ for brevity
- rename Summarizer.py → summarizer.py and JinjaRenderer.py → renderer.py (PascalCase → snake_case)
- collapse the single-file renderers/ sub-package into summarizer/renderer.py
- update all imports, README examples, notebook, and docs/reference.md accordingly
