--- src/overviewpy/datafilesummarizer/__init__.py
@@ -1,3 +1,0 @@
-from overviewpy.datafilesummarizer.Summarizer import Summarizer
-
-__all__ = ["Summarizer"]
