+++ src/overviewpy/summarizer/__init__.py
@@ -1,0 +1,3 @@
+from overviewpy.summarizer.summarizer import Summarizer
+
+__all__ = ["Summarizer"]
